+++ tests/test_join.py
@@ -27,36 +27,33 @@ books = [
 
 class JoinOneTestCase(unittest.TestCase):
     def test_with_dict(self):
-        book_data = copy.deepcopy(books)
-        catalog_books = old_join_one(book_data, catalogs_dict, from_='catalog', as_='catalog_name')
+        catalog_books = old_join_one(books, catalogs_dict, from_='catalog', as_='catalog_name')
         self.assertTrue(all(['catalog_name' in book for book in catalog_books]))
         self.assertEqual('Java', catalog_books[1]['catalog_name'])
+        self.assertFalse('catalog_name' in books[1])
 
     def test_with_choices(self):
-        book_data = copy.deepcopy(books)
-        catalog_books = old_join_one(book_data, catalog_choices, from_='catalog', as_='catalog_name')
+        catalog_books = old_join_one(books, catalog_choices, from_='catalog', as_='catalog_name')
         self.assertTrue(all(['catalog_name' in book for book in catalog_books]))
         self.assertEqual('Java', catalog_books[1]['catalog_name'])
 
     def test_join_one_with_default(self):
-        book_data = copy.deepcopy(books)
         cur_catalogs_dict = {
             1: 'Python',
             2: 'Java'
         }
 
-        catalog_books = old_join_one(book_data, cur_catalogs_dict, from_='catalog', as_='catalog_name')
+        catalog_books = old_join_one(books, cur_catalogs_dict, from_='catalog', as_='catalog_name')
         self.assertTrue(all(['catalog_name' in book for book in catalog_books]))
         self.assertEqual(None, catalog_books[2]['catalog_name'])
 
     def test_join_one_with_custom_default(self):
-        book_data = copy.deepcopy(books)
         cur_catalogs_dict = {
             1: 'Python',
             2: 'Java'
         }
 
-        catalog_books = old_join_one(book_data, cur_catalogs_dict, from_='catalog', as_='catalog_name',
+        catalog_books = old_join_one(books, cur_catalogs_dict, from_='catalog', as_='catalog_name',
                                      default='[未知分类]')
         self.assertTrue(all(['catalog_name' in book for book in catalog_books]))
         self.assertEqual('[未知分类]', catalog_books[2]['catalog_name'])
@@ -64,8 +61,8 @@ class JoinOneTestCase(unittest.TestCase):
 
 class JoinTestCase(unittest.TestCase):
     def test_as_kwargs(self):
-        book_data = copy.deepcopy(books)
-        catalog_books = old_join(book_data, catalogs_list, from_='catalog', to_='id',
+        catalog_books = old_join(books, catalogs_list, from_='catalog', to_='id',
                                  as_kwargs={'name': 'catalog_name'})
         self.assertTrue(all(['catalog_name' in book for book in catalog_books]))
         self.assertEqual('Java', catalog_books[1]['catalog_name'])
+        self.assertFalse('catalog_name' in books[1])